Code segregation, comments, API documentation
