+++ app/layout.tsx
@@ -3,13 +3,12 @@ import Footer from 'components/layout/footer';
 import { Navbar } from 'components/layout/navbar';
 import { WelcomeToast } from 'components/welcome-toast';
 import { GeistSans } from 'geist/font/sans';
-import { getCart, getMenu } from 'lib/sfdc';
+import { getCart, getCategories, Cart } from 'lib/sfdc';
 import { ensureStartsWith } from 'lib/utils';
 import { ReactNode, Suspense } from 'react';
 import { Toaster } from 'sonner';
 import './globals.css';
-import { getCartIdFromCookie, getIsGuestUserFromCookie } from './api/auth/authUtil';
-import { Cart } from 'lib/sfdc/types';
+import { getCartIdFromCookie, getIsGuestUserFromCookie } from './api/auth/cookieUtils';
 import Loading from './loading';
 
 const { TWITTER_CREATOR, TWITTER_SITE, SITE_NAME } = process.env;
@@ -40,12 +39,11 @@ export const metadata = {
 };
 
 export default async function RootLayout({ children }: { children: ReactNode }) {
-  const cartId = await getCartIdFromCookie();
   const isGuestUser = await getIsGuestUserFromCookie();
-  const menuPromise = getMenu();
+  const categoriesPromise = getCategories();
   
   const cartPromise: Promise<Cart | undefined> = isGuestUser !== null 
-    ? getCart(cartId!) 
+    ? getCart() 
     : Promise.resolve(undefined);
 
   return (
@@ -53,7 +51,7 @@ export default async function RootLayout({ children }: { children: ReactNode }) 
       <body className="bg-neutral-50 text-black selection:bg-teal-300 dark:bg-neutral-900 dark:text-white dark:selection:bg-pink-500 dark:selection:text-white min-h-screen flex flex-col">
         <CartProvider cartPromise={cartPromise}>
           <Suspense fallback={<div className="h-16 w-full animate-pulse bg-neutral-100" />}>
-            <Navbar isGuestUser={isGuestUser} menuPromise={menuPromise} />
+            <Navbar isGuestUser={isGuestUser} categoriesPromise={categoriesPromise} />
           </Suspense>
           <main className="flex-grow">
             <Suspense fallback={<Loading />}>
@@ -63,7 +61,7 @@ export default async function RootLayout({ children }: { children: ReactNode }) 
             <WelcomeToast />
           </main>
           <Suspense fallback={<div className="h-16 w-full animate-pulse bg-neutral-100" />}>
-            <Footer menuPromise={menuPromise} />
+            <Footer categoriesPromise={categoriesPromise} />
           </Suspense>
         </CartProvider>
       </body>
