+++ app/search/layout.tsx
@@ -1,10 +1,7 @@
-import Footer from 'components/layout/footer';
 import Collections from 'components/layout/search/collections';
-import FilterList from 'components/layout/search/filter';
-import { sorting } from 'lib/constants';
 import ChildrenWrapper from './children-wrapper';
 import { Suspense } from 'react';
-import Loading from './loading';
+import Loading from './product-grids-loading';
 
 export default function SearchLayout({ children }: { children: React.ReactNode }) {
   return (
