+++ app/sitemap.ts
@@ -1,4 +1,4 @@
-import { getCollections, getPages, getProducts } from 'lib/sfdc';
+import { getCategories, getPages, getProducts } from 'lib/sfdc';
 import { validateEnvironmentVariables } from 'lib/utils';
 import { MetadataRoute } from 'next';
 
@@ -21,10 +21,10 @@ export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
     lastModified: new Date().toISOString()
   }));
 
-  const collectionsPromise = getCollections().then((collections) =>
-    collections.map((collection) => ({
-      url: `${baseUrl}${collection.path}`,
-      lastModified: collection.updatedAt
+  const categoriesPromise = getCategories().then((categories) =>
+    categories.map((category) => ({
+      url: `${baseUrl}${category.path}`,
+      lastModified: category.updatedAt
     }))
   );
 
@@ -43,9 +43,8 @@ export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
   );
 
   let fetchedRoutes: Route[] = [];
-
   try {
-    fetchedRoutes = (await Promise.all([collectionsPromise, productsPromise, pagesPromise])).flat();
+    fetchedRoutes = (await Promise.all([categoriesPromise, productsPromise, pagesPromise])).flat() as Route[];
   } catch (error) {
     throw JSON.stringify(error, null, 2);
   }
