+++ components/product/variant-selector.tsx
@@ -2,7 +2,7 @@
 
 import clsx from 'clsx';
 import { useProduct, useUpdateURL } from 'components/product/product-context';
-import { ProductOption, ProductVariant } from 'lib/sfdc/types';
+import { ProductOption, ProductVariant } from 'lib/sfdc';
 
 type Combination = {
   id: string;
