+++ lib/sfdc/pageApis.ts
@@ -1,0 +1,18 @@
+import { Page } from './types';
+
+/**
+ * Fetches a single page by its handle or ID from the SFDC API.
+ * @param {string} handle - The page handle or ID.
+ * @returns {Promise<Page | undefined>} The page object, or undefined if not found.
+ */
+export async function getPage(handle: string): Promise<Page | undefined> {
+  return undefined;
+}
+
+/**
+ * Fetches all pages from the SFDC API.
+ * @returns {Promise<Page[]>} An array of pages.
+ */
+export async function getPages(): Promise<Page[]> {
+  return [];
+}
