+++ lib/utils.ts
@@ -11,7 +11,13 @@ export const ensureStartsWith = (stringToCheck: string, startsWith: string) =>
   stringToCheck.startsWith(startsWith) ? stringToCheck : `${startsWith}${stringToCheck}`;
 
 export const validateEnvironmentVariables = () => {
-  const requiredEnvironmentVariables = ['SHOPIFY_STORE_DOMAIN', 'SHOPIFY_STOREFRONT_ACCESS_TOKEN'];
+  const requiredEnvironmentVariables = [
+    'SFDC_COMMERCE_API_VERSION',
+    'SFDC_COMMERCE_WEBSTORE_ID',
+    'SFDC_COMMERCE_WEBSTORE_NAME',
+    'SFDC_COMMERCE_WEBSTORE_SITE_ID',
+    'SFDC_COMMERCE_WEBSTORE_SITE_URL'
+  ];
   const missingEnvironmentVariables = [] as string[];
 
   requiredEnvironmentVariables.forEach((envVar) => {
@@ -22,18 +28,16 @@ export const validateEnvironmentVariables = () => {
 
   if (missingEnvironmentVariables.length) {
     throw new Error(
-      `The following environment variables are missing. Your site will not work without them. Read more: https://vercel.com/docs/integrations/shopify#configure-environment-variables\n\n${missingEnvironmentVariables.join(
-        '\n'
-      )}\n`
+      `The following environment variables are missing. Your site will not work without them.\n\n${missingEnvironmentVariables.join('\n')}\n`
     );
   }
 
   if (
-    process.env.SHOPIFY_STORE_DOMAIN?.includes('[') ||
-    process.env.SHOPIFY_STORE_DOMAIN?.includes(']')
+    process.env.SFDC_COMMERCE_WEBSTORE_SITE_URL?.includes('[') ||
+    process.env.SFDC_COMMERCE_WEBSTORE_SITE_URL?.includes(']')
   ) {
     throw new Error(
-      'Your `SHOPIFY_STORE_DOMAIN` environment variable includes brackets (ie. `[` and / or `]`). Your site will not work with them there. Please remove them.'
+      'Your `SFDC_COMMERCE_WEBSTORE_SITE_URL` environment variable includes brackets (ie. `[` and / or `]`). Your site will not work with them there. Please remove them.'
     );
   }
 };
